Time: 0 ms (100.00%) | Memory: 7.8 MB (52.58%) - LeetSync

+++ 1064-smallest-integer-divisible-by-k/smallest-integer-divisible-by-k.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int smallestRepunitDivByK(int k) {
         if(k % 2 == 0 || k % 5 == 0) return -1;
-        int rem = 0;
+        long long rem = 0;
         for(int n = 1; n<=k; n++){
             rem = (rem * 10 + 1) % k;
             if(rem == 0) return n;
